# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * MedScribe AI API — transcription, clinical structuring and per-physician storage of consultations. Every endpoint except /healthz and /templates requires a Supabase access token and only ever sees the caller's own rows.
 * OpenAPI spec version: 0.1.0
 */
export interface RuntimeConfig {
  supabaseUrl: string;
  /** Publishable (anon) key — safe to expose, gated by RLS */
  supabasePublishableKey: string;
}

export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export type TemplateId = typeof TemplateId[keyof typeof TemplateId];


export const TemplateId = {
  soap: 'soap',
  evolucion: 'evolucion',
  informe: 'informe',
  receta: 'receta',
} as const;

export interface Template {
  id: TemplateId;
  /** Short display name in Spanish (e.g. "SOAP") */
  name: string;
  /** One-line description of when to use it */
  description: string;
  /** Ordered section labels the AI will produce for this template */
  sections: string[];
}

export interface AnonymizeInput {
  /** @minLength 1 */
  text: string;
}

export interface AnonymizeResult {
  text: string;
  /** Number of identifiers replaced */
  replacements: number;
}

export interface NoteInput {
  /**
     * Raw transcript or dictation
     * @minLength 1
     */
  text: string;
  template: TemplateId;
  /** Run the local anonymizer before sending the text to the AI */
  anonymize?: boolean;
  /** Standing style instructions the physician wants applied to every note */
  preferences?: string[];
}

export interface NoteSection {
  label: string;
  /** Plain text, no markdown */
  content: string;
}

export interface StructuredNote {
  template: TemplateId;
  /** Short auto-generated title (e.g. chief complaint) for the history list */
  title: string;
  sections: NoteSection[];
  /** Full note as clean plain text, ready to paste into any EHR */
  plainText: string;
  anonymized?: boolean;
  processedAt: string;
}

export interface RefineInput {
  note: StructuredNote;
  /**
     * What to change in this note
     * @minLength 1
     */
  instruction: string;
  /** Standing style instructions applied to every note */
  preferences?: string[];
}

/**
 * Keep the recording in the private bucket and return its object key
 */
export type AudioUploadStore = typeof AudioUploadStore[keyof typeof AudioUploadStore];


export const AudioUploadStore = {
  true: 'true',
  false: 'false',
} as const;

export interface AudioUpload {
  file: Blob;
  /** ISO language hint, defaults to "es" */
  language?: string;
  /** Keep the recording in the private bucket and return its object key */
  store?: AudioUploadStore;
}

export interface Transcription {
  text: string;
  durationSeconds: number;
  /** Object key of the stored recording, when store was requested */
  audioPath?: string | null;
}

export interface Account {
  id: string;
  email: string;
  fullName?: string | null;
}

export interface Preferences {
  /** Standing style instructions applied to every note */
  preferences: string[];
}

export interface DeletedCount {
  deleted: number;
}

export interface AudioLink {
  url: string;
  /** Seconds the signed URL stays valid */
  expiresIn: number;
}

export interface Consultation {
  id: string;
  /** Free-text label chosen by the physician to recognise the consultation */
  patientRef?: string | null;
  template: TemplateId;
  title: string;
  sections: NoteSection[];
  plainText: string;
  anonymized: boolean;
  /** Raw transcript the note was generated from */
  transcript?: string | null;
  /** Object key of the recording in the private storage bucket */
  audioPath?: string | null;
  audioDurationSeconds?: number | null;
  createdAt: string;
  updatedAt: string;
}

export interface ConsultationInput {
  note: StructuredNote;
  patientRef?: string;
  transcript?: string;
  audioPath?: string;
  audioDurationSeconds?: number;
}

export interface ConsultationUpdate {
  note?: StructuredNote;
  patientRef?: string;
}

export type ListConsultationsParams = {
/**
 * @minimum 1
 * @maximum 200
 */
limit?: number;
};