# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: >-
    MedScribe AI API — transcription, clinical structuring and per-physician
    storage of consultations. Every endpoint except /healthz and /templates
    requires a Supabase access token and only ever sees the caller's own rows.
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /audio/transcribe:
    post:
      operationId: transcribeAudio
      tags: [audio]
      summary: Transcribe an audio recording or uploaded file (webm, mp3, m4a, wav, ogg) to text
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/AudioUpload"
      responses:
        "200":
          description: Transcript
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transcription"
        "400":
          description: Invalid audio
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /me:
    get:
      operationId: getCurrentUser
      tags: [account]
      summary: Profile of the authenticated physician
      responses:
        "200":
          description: Account
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Account"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /preferences:
    get:
      operationId: getPreferences
      tags: [preferences]
      summary: Standing style preferences of the authenticated physician
      responses:
        "200":
          description: Preferences
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Preferences"
    put:
      operationId: setPreferences
      tags: [preferences]
      summary: Replace the standing style preferences
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Preferences"
      responses:
        "200":
          description: Saved preferences
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Preferences"
  /consultations:
    get:
      operationId: listConsultations
      tags: [consultations]
      summary: List the authenticated physician's consultations, newest first
      parameters:
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 200
            default: 50
      responses:
        "200":
          description: Consultations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Consultation"
    post:
      operationId: createConsultation
      tags: [consultations]
      summary: Store a structured note as a consultation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConsultationInput"
      responses:
        "201":
          description: Stored consultation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Consultation"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteAllConsultations
      tags: [consultations]
      summary: Delete every consultation and stored recording of the authenticated physician
      responses:
        "200":
          description: Number of consultations deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeletedCount"
  /consultations/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    get:
      operationId: getConsultation
      tags: [consultations]
      summary: Fetch one consultation
      responses:
        "200":
          description: Consultation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Consultation"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateConsultation
      tags: [consultations]
      summary: Update a stored consultation, e.g. after refining the note
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConsultationUpdate"
      responses:
        "200":
          description: Updated consultation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Consultation"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteConsultation
      tags: [consultations]
      summary: Delete one consultation and its stored recording
      responses:
        "200":
          description: Number of consultations deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeletedCount"
  /consultations/{id}/audio:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    get:
      operationId: getConsultationAudio
      tags: [consultations]
      summary: Short-lived signed URL for the consultation's stored recording
      responses:
        "200":
          description: Signed URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AudioLink"
        "404":
          description: No recording stored for this consultation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Supabase access token of the signed-in physician
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
    TemplateId:
      type: string
      enum: [soap, evolucion, informe, receta]
...
    NoteSection:
      type: object
      required: [label, content]
      properties:
        label:
          type: string
        content:
          type: string
          description: Plain text, no markdown
    StructuredNote:
      type: object
      required: [template, title, sections, plainText, processedAt]
      properties:
        template:
          $ref: "#/components/schemas/TemplateId"
        title:
          type: string
          description: Short auto-generated title (e.g. chief complaint) for the history list
        sections:
          type: array
          items:
            $ref: "#/components/schemas/NoteSection"
        plainText:
          type: string
          description: Full note as clean plain text, ready to paste into any EHR
        anonymized:
          type: boolean
        processedAt:
          type: string
          format: date-time
    AudioUpload:
      type: object
      required: [file]
      properties:
        file:
          type: string
          format: binary
        language:
          type: string
          description: ISO language hint, defaults to "es"
        store:
          type: string
          enum: ["true", "false"]
          description: Keep the recording in the private bucket and return its object key
    Transcription:
      type: object
      required: [text, durationSeconds]
      properties:
        text:
          type: string
        durationSeconds:
          type: number
        audioPath:
          type: string
          nullable: true
          description: Object key of the stored recording, when store was requested
    Account:
      type: object
      required: [id, email]
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
        fullName:
          type: string
          nullable: true
    Preferences:
      type: object
      required: [preferences]
      properties:
        preferences:
          type: array
          description: Standing style instructions applied to every note
          items:
            type: string
    DeletedCount:
      type: object
      required: [deleted]
      properties:
        deleted:
          type: integer
    AudioLink:
      type: object
      required: [url, expiresIn]
      properties:
        url:
          type: string
        expiresIn:
          type: integer
          description: Seconds the signed URL stays valid
    Consultation:
      type: object
      required:
        [id, template, title, sections, plainText, anonymized, createdAt, updatedAt]
      properties:
        id:
          type: string
          format: uuid
        patientRef:
          type: string
          nullable: true
          description: Free-text label chosen by the physician to recognise the consultation
        template:
          $ref: "#/components/schemas/TemplateId"
        title:
          type: string
        sections:
          type: array
          items:
            $ref: "#/components/schemas/NoteSection"
        plainText:
          type: string
        anonymized:
          type: boolean
        transcript:
          type: string
          nullable: true
          description: Raw transcript the note was generated from
        audioPath:
          type: string
          nullable: true
          description: Object key of the recording in the private storage bucket
        audioDurationSeconds:
          type: number
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    ConsultationInput:
      type: object
      required: [note]
      properties:
        note:
          $ref: "#/components/schemas/StructuredNote"
        patientRef:
          type: string
        transcript:
          type: string
        audioPath:
          type: string
        audioDurationSeconds:
          type: number
    ConsultationUpdate:
      type: object
      properties:
        note:
          $ref: "#/components/schemas/StructuredNote"
        patientRef:
          type: string