# lib/api-zod/src/generated/types/account.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * MedScribe AI API — transcription, clinical structuring and per-physician storage of consultations. Every endpoint except /healthz and /templates requires a Supabase access token and only ever sees the caller's own rows.
 * OpenAPI spec version: 0.1.0
 */

export interface Account {
  id: string;
  email: string;
  fullName?: string | null;
}