# lib/api-zod/src/generated/types/consultation.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * MedScribe AI API — transcription, clinical structuring and per-physician storage of consultations. Every endpoint except /healthz and /templates requires a Supabase access token and only ever sees the caller's own rows.
 * OpenAPI spec version: 0.1.0
 */
import type { NoteSection } from './noteSection';
import type { TemplateId } from './templateId';

export interface Consultation {
  id: string;
  /** Free-text label chosen by the physician to recognise the consultation */
  patientRef?: string | null;
  template: TemplateId;
  title: string;
  sections: NoteSection[];
  plainText: string;
  anonymized: boolean;
  /** Raw transcript the note was generated from */
  transcript?: string | null;
  /** Object key of the recording in the private storage bucket */
  audioPath?: string | null;
  audioDurationSeconds?: number | null;
  createdAt: Date;
  updatedAt: Date;
}